chore: move preflight apis to externalapis dir
