+++ externalapis/preflight/v1beta2/hostpreflight_types.go
