+++ internal/preflight/analyzer/access_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/replicatedhq/troubleshoot/pkg/collect"
 	"k8s.io/apimachinery/pkg/version"
 
-	preflightv1beta2 "github.com/apecloud/kubeblocks/apis/preflight/v1beta2"
+	preflightv1beta2 "github.com/apecloud/kubeblocks/externalapis/preflight/v1beta2"
 )
 
 var _ = Describe("access_test", func() {
