+++ internal/preflight/analyzer/anzlyzer_test.go
@@ -29,7 +29,7 @@ import (
 	"github.com/replicatedhq/troubleshoot/pkg/collect"
 	"k8s.io/apimachinery/pkg/version"
 
-	preflightv1beta2 "github.com/apecloud/kubeblocks/apis/preflight/v1beta2"
+	preflightv1beta2 "github.com/apecloud/kubeblocks/externalapis/preflight/v1beta2"
 	kbcollector "github.com/apecloud/kubeblocks/internal/preflight/collector"
 )
