+++ internal/preflight/collector/host_utility_test.go
@@ -26,7 +26,7 @@ import (
 
 	troubleshoot "github.com/replicatedhq/troubleshoot/pkg/apis/troubleshoot/v1beta2"
 
-	preflightv1beta2 "github.com/apecloud/kubeblocks/apis/preflight/v1beta2"
+	preflightv1beta2 "github.com/apecloud/kubeblocks/externalapis/preflight/v1beta2"
 )
 
 var _ = Describe("host_utility_test", func() {
