+++ internal/preflight/concat_spec.go
@@ -19,7 +19,7 @@ package preflight
 import (
 	troubleshoot "github.com/replicatedhq/troubleshoot/pkg/apis/troubleshoot/v1beta2"
 
-	preflightv1beta2 "github.com/apecloud/kubeblocks/apis/preflight/v1beta2"
+	preflightv1beta2 "github.com/apecloud/kubeblocks/externalapis/preflight/v1beta2"
 )
 
 // ConcatPreflightSpec splices multiple PreflightSpec into one Preflight object
